requirements.txt
streamlit
Pillow
numpy
matplotlib
torch
torchvision
opencv-python-headless
